@@ -1,35 +1,34 @@
 [tool.poetry]
-name = "security_guard"
+name = "crazy_bear"
 version = "0.1.0"
 description = ""
 authors = ["expert_m"]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.11"
-requests = "^2.23"
-pytz = "^2022.1"
-emoji = "^2.0.0"
-pyserial = "^3.4"
-#opencv-python = { version = "^4.2.0", optional = true }
-imutils = "^0.5.3"
-numpy = { version = "1.21.5", optional = true }
-matplotlib = { version = "=3.5.1", optional = true }
-sqlalchemy = "^1.3.23"
-sentry-sdk = "^1.0.0"
-dropbox = "^11.4.0"
-openpyxl = "^3.0.3"
-pytest = "^7.0.1"
-wheel = "^0.37.0"
-pandas = { version = "1.3.5", optional = true }
-scipy = { version = "1.8.0", optional = true }
-python-dotenv = "^0.20.0"
-python-telegram-bot = "^13.8.1"
-crontab = "^0.23.0"
-psycopg2-binary = { version = "=2.9.2", optional = true }
-pika = "^1.2.0"
+python = ">=3.10,<3.12"
+requests = "^2.31.0"
+pytz = "^2023.3.post1"
+emoji = "^2.8.0"
+pyserial = "^3.5"
+opencv-python = { version = "~4.6.0", optional = true }
+imutils = "^0.5.4"
+numpy = { version = "~1.24.2", optional = true }
+matplotlib = { version = "~3.6.3", optional = true }
+sqlalchemy = {extras = ["asyncio"], version = "^2.0.21"}
+sentry-sdk = "^1.31.0"
+dropbox = "^11.36.2"
+openpyxl = "^3.1.2"
+pytest = "^7.4.2"
+wheel = "^0.41.2"
+pandas = { version = "~1.5.3", optional = true }
+scipy = { version = "~1.10.1", optional = true }
+python-dotenv = "^1.0.0"
+crontab = "^1.0.1"
+psycopg2-binary = { version = "~2.9.6", optional = true }
 paho-mqtt = "^1.6.1"
-mplcyberpunk = "^0.5.2"
-doc = "^0.1.0"
+mplcyberpunk = "^0.7.0"
+python-telegram-bot = "^20.6"
+pika = "^1.3.2"
 
 [tool.poetry.dev-dependencies]
 
