@@ -55,7 +55,7 @@ class BaseSignalHandler(abc.ABC):
         pass
 
     def generate_plots(self, *,
-                       date_range: typing.Tuple[datetime.datetime, datetime.datetime],
+                       date_range: tuple[datetime.datetime, datetime.datetime],
                        components: typing.Set[str]) -> typing.Optional[typing.Sequence[io.BytesIO]]:
         return None
 
