@@ -33,7 +33,7 @@ class CpuTempHandler(BaseAdvancedSignalHandler):
             return NOTHING
 
     def generate_plots(self, *,
-                       date_range: typing.Tuple[datetime.datetime, datetime.datetime],
+                       date_range: tuple[datetime.datetime, datetime.datetime],
                        components: typing.Set[str]) -> typing.Optional[typing.Sequence[io.BytesIO]]:
         if 'inner_stats' not in components:
             return None
