@@ -88,7 +88,7 @@ class RouterHandler(BaseAdvancedSignalHandler):
             self._state[constants.USER_IS_CONNECTED_TO_ROUTER] = False
 
     def generate_plots(self, *,
-                       date_range: typing.Tuple[datetime.datetime, datetime.datetime],
+                       date_range: tuple[datetime.datetime, datetime.datetime],
                        components: typing.Set[str]) -> typing.Optional[typing.Sequence[io.BytesIO]]:
 
         if 'router_usage' not in components:
